requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.25.2
pillow==10.1.0
google-generativeai==0.3.2
apscheduler==3.10.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
